requirements.txt
pyserial==2.6
PyDispatcher==2.0.3
bottle==0.11.6
Sphinx==1.1.3
